ui - refactor ui_get_spend_by_tags
